# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Fem-ho
  version: 0.1.0
  description: |
    Contracte de l'API de Fem-ho. Aquest fitxer és la font de veritat (regla 5
    d'instruccions.md): els tipus de TypeScript i el client de Kotlin es generen des
    d'aquí i cap client escriu tipus a mà. Un endpoint que no hi és no existeix.

    A M1 només hi ha les dues rutes públiques de vida i identitat. La resta arriba a
    la seva fita, seguint docs/05-api-rest.md.
  license:
    name: AGPL-3.0-or-later
    identifier: AGPL-3.0-or-later
servers:
  - url: /api/v1
    description: Base de l'API. GET /info i GET /healthz pengen de l'arrel, no d'aquí.
...
paths:
...
  /tasks/{id}/checklists:
    get:
      tags: [checklists]
      operationId: listChecklists
      summary: Les llistes senzilles d'una tasca
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Llistes amb els seus ítems.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Checklist'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [checklists]
      operationId: createChecklist
      summary: Crear una llista dins d'una tasca
      description: |
        Una llista pertany **sempre** a una tasca, i opcionalment s'ancora a una subtasca
        concreta (P1). No és un subprojecte ni una subtasca: són dues taules i un flag de
        presentació.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChecklistInput'
      responses:
        '201':
          description: Creada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Checklist'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /checklist-items/{id}:
    patch:
      tags: [checklists]
      operationId: updateChecklistItem
      summary: Marcar o desmarcar un ítem
      description: |
        **Marcar-lo pot disparar la cascada amunt** (docs/01 §4): quan l'últim ítem d'una
        llista passa a fet, en la mateixa transacció es marca la subtasca ancorada; i si
        totes les llistes i subtasques d'una tasca estan fetes, la tasca passa a `done`.

        Es registra amb `verb='cascade_complete'` perquè es distingeixi d'un gest directe.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChecklistItemInput'
      responses:
        '200':
          description: Modificat, amb el que la cascada hagi tocat.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChecklistItemResult'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [checklists]
      operationId: deleteChecklistItem
      summary: Esborrar un ítem
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        '204': { description: Esborrat. }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /shares:
    get:
      tags: [shares]
      operationId: listShares
      summary: Els enllaços públics de qui pregunta
      description: |
        Es llisten amb la configuració i les visites, **mai amb el token**: només se'n
        guarda l'HMAC (docs/10 §3). Si l'usuari el perd, ha de crear-ne un de nou.
      responses:
        '200':
          description: Els enllaços.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/ShareSummary' }
        '401':
          $ref: '#/components/responses/Unauthenticated'
    post:
      tags: [shares]
      operationId: createShare
      summary: Crear un enllaç públic
      description: |
        Una **tasca amb les seves llistes**, o una **llista sola**. No es comparteixen
        projectes ni àmbits sencers (docs/10 §1).

        **No hi ha permís d'edició** (D10): un convidat marca ítems i comenta.

        L'URL sencer es retorna **un sol cop**.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                task_id: { type: string }
                checklist_id: { type: string }
                permission:
                  type: string
                  enum: [view, check, comment]
                  default: view
                password: { type: [string, 'null'] }
                require_name: { type: boolean, default: false }
                expires_at: { type: [string, 'null'], format: date-time }
                max_views: { type: [integer, 'null'] }
      responses:
        '201':
          description: Creat. Porta l'URL, i és l'única vegada.
          content:
            application/json:
              schema:
                type: object
                required: [url, token, share]
                properties:
                  url: { type: string }
                  token: { type: string }
                  share: { $ref: '#/components/schemas/ShareSummary' }
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
  /shares/{id}:
    delete:
      tags: [shares]
      operationId: revokeShare
      summary: Revocar un enllaç
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '204':
          description: Revocat.
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '404':
          $ref: '#/components/responses/NotFound'

    get:
      tags: [scopes]
      operationId: getShare
      summary: Un enllaç compartit
      description: "El token NO hi surt: no es pot recuperar del seu HMAC (docs/10 §6)."
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        '200':
          description: L'enllaç.
          content: { application/json: { schema: { $ref: '#/components/schemas/ShareSummary' } } }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [scopes]
      operationId: updateShare
      summary: Reconfigurar un enllaç
      description: |
        El token NO es regenera mai des d'aquí: si canviar el permís canviés l'enllaç, tot
        el que ja s'hagi enviat deixaria de funcionar sense que ningú ho hagi demanat.

        Canviar la contrasenya desbloqueja l'enllaç: els cinc intents fallits eren contra
        l'antiga.
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                permission: { type: string, enum: [view, check, comment] }
                expires_at: { type: [string, 'null'], format: date-time }
                max_views: { type: [integer, 'null'] }
                require_name: { type: boolean }
                password:
                  type: [string, 'null']
                  description: '`null` la treu; una cadena la canvia; absent no la toca.'
      responses:
        '200':
          description: Reconfigurat.
          content: { application/json: { schema: { $ref: '#/components/schemas/ShareSummary' } } }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '409':
          description: Està revocat i no es pot tornar a obrir.
          content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
...
  /tasks/{id}:
    get:
      tags: [tasks]
      operationId: getTask
      summary: Una tasca
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        '200':
          description: La tasca.
          content: { application/json: { schema: { $ref: '#/components/schemas/Task' } } }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [tasks]
      operationId: updateTask
      summary: Modificar una tasca
      description: |
        Només els camps que es donin. `null` vol dir "buida'l" i absent vol dir "no el
        toquis": sense la distinció, buidar una data seria impossible des d'un client
        que envia només el que ha canviat.
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/TaskPatch' } } }
      responses:
        '200':
          description: Modificada.
          content: { application/json: { schema: { $ref: '#/components/schemas/Task' } } }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [tasks]
      operationId: deleteTask
      summary: Esborrar una tasca
      description: |
        Esborrat suau. Les subtasques i les llistes cauen amb ella —no existeixen fora
        d'una tasca—, a diferència del que passa en esborrar un projecte.
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        '204': { description: Esborrada. }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /tasks/{id}/comments:
    get:
      tags: [tasks]
      operationId: listComments
      summary: Els comentaris d'una tasca
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        '200':
          description: Els comentaris.
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Comment' } }
        '401': { $ref: '#/components/responses/Unauthenticated' }
    post:
      tags: [tasks]
      operationId: addComment
      summary: Comentar una tasca
      description: |
        És la via principal perquè un agent reporti (docs/09 §6), i per això surt a
        l'historial com qualsevol altre gest.
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [body]
              properties:
                body: { type: string }
      responses:
        '201':
          description: Comentat.
          content: { application/json: { schema: { $ref: '#/components/schemas/Comment' } } }
        '401': { $ref: '#/components/responses/Unauthenticated' }
...
  /shares/{id}/accesses:
    get:
      tags: [scopes]
      operationId: listShareAccesses
      summary: Qui ha entrat per un enllaç
      description: |
        L'etiqueta és pseudònima: "Extern · Marta" o "Extern · a4f2". No hi ha cap
        columna d'IP enlloc, i és una decisió de privadesa explícita (D10).
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        '200':
          description: Els accessos.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/ShareAccess' } }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  schemas:
...
    TaskPatch:
      type: object
      description: |
        `null` vol dir "buida'l" i absent vol dir "no el toquis". Per moure-la, `/move`.
      properties:
        title: { type: string }
        description: { type: [string, 'null'] }
        due_date: { type: [string, 'null'], format: date }
        due_time: { type: [string, 'null'] }
        deadline:
          type: [string, 'null']
          format: date-time
          description: |
            Separada del venciment (docs/02 §7): "fes-ho aquest dijous" i "com a molt
            tard el dia 30" són dues coses, i amb un sol camp qui té les dues n'ha de
            triar una.
        project_id:
          type: [string, 'null']
          description: |
            `null` la torna a l'espai general de l'àmbit. **L'àmbit no es pot canviar
            editant**: altres membres, altres etiquetes i l'assignació automàtica dels
            àmbits individuals. Un projecte d'un altre àmbit es rebutja amb `422`.
        rrule:
          type: [string, 'null']
          description: 'RFC 5545. `null` deixa de repetir-se.'
        recurrence_mode:
          type: string
          enum: [schedule, completion]
          description: |
            La distinció que Todoist escriu com a `every` contra `every!`: repetir-se
            cada dimarts contra repetir-se una setmana **després d'haver-la fet**. RRULE
            no sap expressar la segona, i per a tasques domèstiques és la que la gent
            vol; per això és un camp propi (docs/01 §4).
        ai_mode: { type: string, enum: [manual, assisted, delegated] }
        ai_instructions: { type: [string, 'null'] }
...
    ShareSummary:
      type: object
      required: [id, permission, require_name, has_password, view_count, created_at]
      properties:
        id: { type: string }
        task_id: { type: [string, 'null'] }
        checklist_id: { type: [string, 'null'] }
        permission:
          type: string
          enum: [view, check, comment]
        require_name: { type: boolean }
        has_password:
          type: boolean
          description: Si en té, no quina. La contrasenya es guarda amb argon2id.
        expires_at: { type: [string, 'null'], format: date-time }
        max_views: { type: [integer, 'null'] }
        view_count: { type: integer }
        created_at: { type: string, format: date-time }
        revoked_at: { type: [string, 'null'], format: date-time }
...
    Checklist:
      type: object
      additionalProperties: false
      required: [id, task_id, name, pinned, show_completed_inline, position, items, version]
      properties:
        id: { type: string }
        task_id: { type: string }
        subtask_id:
          type: [string, 'null']
          description: Ancoratge opcional a una subtasca concreta (P1).
        name: { type: string }
        pinned:
          type: boolean
          description: Pinejat PER USUARI. El rail és de cada persona.
        show_completed_inline:
          type: boolean
          description: |
            Completats ratllats al seu lloc, o en una secció "Completats" al final,
            plegada amb el recompte (docs/02 §6).
        position: { type: string }
        items:
          type: array
          items:
            $ref: '#/components/schemas/ChecklistItem'
        version: { type: integer }
    ChecklistInput:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        id: { type: string }
        name: { type: string, minLength: 1 }
        subtask_id: { type: string }
        show_completed_inline: { type: boolean }
...
    ChecklistItemInput:
      type: object
      additionalProperties: false
      properties:
        id: { type: string }
        text: { type: string, minLength: 1 }
        done: { type: boolean }
        position: { type: string }
    ChecklistItemResult:
      type: object
      additionalProperties: false
      required: [item, cascade]
      properties:
        item:
          $ref: '#/components/schemas/ChecklistItem'
        cascade:
          type: object
          additionalProperties: false
          required: [checklist_completed, subtask_completed, task_completed]
          description: Què ha tocat la cascada amunt, perquè la interfície ho pugui dir.
          properties:
            checklist_completed: { type: boolean }
            subtask_completed: { type: boolean }
            task_completed: { type: boolean }
            suggest_unpin:
              type: boolean
              description: |
                Cert quan una llista PINEJADA s'ha completat del tot: llavors es pregunta
                si es vol despinejar (P1).
...
    Task:
      type: object
      additionalProperties: false
      required:
        [id, scope_id, title, status, position, ai_mode, view_mode, created_at, updated_at, version]
      properties:
        id: { type: string }
        scope_id: { type: string }
        project_id: { type: [string, 'null'] }
        title: { type: string }
        description: { type: [string, 'null'] }
        status:
          type: string
          enum: [inbox, todo, doing, done]
          description: La columna. Es diu `status`, no `column` (D2).
        position:
          type: string
          description: Índex fraccional. Ordenat amb collation binària (D3).
        due_date:
          type: [string, 'null']
          description: YYYY-MM-DD, sense fus. Una data de tot el dia no té instant.
        due_time: { type: [string, 'null'] }
        deadline: { type: [string, 'null'], format: date-time }
        completed_at: { type: [string, 'null'], format: date-time }
        view_mode:
          type: string
          enum: [card, simple]
        ai_mode:
          type: string
          enum: [manual, assisted, delegated]
          description: Delegar no és assignar (D5).
        delegate_agent_id: { type: [string, 'null'] }
        rrule:
          type: [string, 'null']
          description: 'RFC 5545, o `null` si no es repeteix.'
        recurrence_mode:
          type: [string, 'null']
          enum: [schedule, completion, null]
          description: |
            `schedule` compta des del venciment; `completion`, des que es fa. És la
            distinció que Todoist escriu com a `every` contra `every!` i que RRULE no sap
            expressar (docs/01 §4).
        recurrence_parent_id:
          type: [string, 'null']
          description: La instància de la qual ve, si n'és una repetició.
        progress:
          type: object
          required: [done, total, lists]
          description: |
            Subtasques i ítems de llista **comptats junts**. La targeta els ensenya sota
            un sol commutador ("3/7 fets") perquè per a qui mira el tauler són el mateix:
            coses que falten dins d'aquesta tasca. La distinció importa en obrir-la.

            És un agregat i no la llista: portar tots els ítems de totes les targetes
            faria baixar milers de files per pintar unes quantes pastilles.
          properties:
            done: { type: integer }
            total: { type: integer }
            lists:
              type: integer
              description: |
                Quants blocs desplegables té la targeta: un per les subtasques, si en té,
                i un per cada llista. És el número del commutador "Llistes (2)", que
                compta blocs i no ítems.
        assignee_ids:
          type: array
          items: { type: string }
          description: Persones. El brief demana "persona o persones".
        created_by: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        version: { type: integer }
...
    Problem:
      type: object
      description: |
        Error en application/problem+json (RFC 9457, docs/05 §3).

        **Qui llegeix un error no és qui el pateix.** El reben tres públics molt
        diferents: una persona davant de l'app, un agent per MCP, i un client CalDAV com
        DAVx⁵ o Thunderbird. Els dos últims no tenen ni catàleg ni idioma. Per això
        l'error viatja en dues peces:

        - `type`, `title` i `detail` **en anglès**, estables, per programar-hi.
        - `params`, les **dades** de l'error. Les apps hi posen el text del seu catàleg
          amb la clau `error.<slug>`, on el `slug` és l'últim tram del `type`. Si no en
          tenen la clau, ensenyen el `detail`: així un error nou del servidor no deixa
          mai una pantalla muda ni obliga a desplegar les dues bandes alhora.
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
          examples: ['https://femho.app/errors/scope-forbidden']
        title:
          type: string
          examples: ['Scope not accessible']
        params:
          type: object
          additionalProperties: { type: [string, integer] }
          description: |
            Les dades de l'error, planes. Plans a posta: el client no ha de saber
            recórrer un JSON dins d'un JSON per pintar un missatge.
          examples: [{ name: 'Família', count: 3 }]
        status:
          type: integer
          examples: [403]
        detail:
          type: string
          examples: ["Aquest token només té accés a l'àmbit Feina."]
        instance:
          type: string
  responses:
    Unauthenticated:
      description: Cal autenticar-se.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: |
        Sense permís. El cos diu QUINS àmbits veu el token i ON és el que s'ha demanat:
        un 403 mut fa que un agent reintenti en bucle (docs/05 §2).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: No existeix, o no és visible per a qui pregunta.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: femho_session
      description: Sessió de navegador. Galeta HttpOnly, Secure, SameSite=Lax.
    bearer:
      type: http
      scheme: bearer
      description: |
        Token d'accés (Android) o token d'API `femho_pat_…` (integracions, MCP, IA).
        Les capacitats i els àmbits van al registre del token, no a scopes d'OAuth
        (regla 9).
security:
  - session: []
  - bearer: []